requirements.txt
streamlit
pandas
pymupdf
python-docx